Fix for 'slow' systems (#75)
* On 'slow' systems we need to wait for 'Buy now'

IMO we need to wait until the button displays 'Buy now'. There is a period of time (which I observed could be substantial) where it says "Loading ...". If we populate the amount before it says 'Buy now' the edit box reverts to empty string

The way this check is done is probably not optimal, but I am not familiar with the framework